:recycle: Move the Route Configuration Into a Dedicated File
:art: Load the Global Route Configuration from the Configuration File
